+++ src/Service/Server/Validator/Collection/ServiceServerRequestValidatorCollectionInterface.php
@@ -15,24 +15,19 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2020 (original work) Open Assessment Technologies SA;
+ * Copyright (c) 2021 (original work) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
 
-namespace OAT\Library\Lti1p3Core\Util\Result;
+namespace OAT\Library\Lti1p3Core\Service\Server\Validator\Collection;
 
-interface ResultInterface
+use OAT\Library\Lti1p3Core\Service\Server\Validator\ServiceServerRequestValidatorInterface;
+
+interface ServiceServerRequestValidatorCollectionInterface
 {
-    public function hasError(): bool;
+    public function add(ServiceServerRequestValidatorInterface $validator): ServiceServerRequestValidatorCollectionInterface;
 
-    public function addSuccess(string $success): ResultInterface;
-
-    public function getSuccesses(): array;
-
-    public function setSuccesses(array $successes = []): ResultInterface;
-
-    public function getError(): ?string;
-
-    public function setError(?string $error): ResultInterface;
+    /** @return ServiceServerRequestValidatorInterface[] */
+    public function all(): array;
 }